@@ -8,17 +8,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.3.10",
-    "@angular/cdk": "^17.1.0",
-    "@angular/common": "17.3.10",
-    "@angular/compiler": "17.3.10",
-    "@angular/core": "17.3.10",
-    "@angular/forms": "17.3.10",
-    "@angular/material": "^17.1.0",
-    "@angular/platform-browser": "17.3.10",
-    "@angular/platform-browser-dynamic": "17.3.10",
-    "@angular/router": "17.3.10",
-    "@angular/service-worker": "17.3.10",
+    "@angular/animations": "18.0.1",
+    "@angular/cdk": "^18.0.1",
+    "@angular/common": "18.0.1",
+    "@angular/compiler": "18.0.1",
+    "@angular/core": "18.0.1",
+    "@angular/forms": "18.0.1",
+    "@angular/material": "^18.0.1",
+    "@angular/platform-browser": "18.0.1",
+    "@angular/platform-browser-dynamic": "18.0.1",
+    "@angular/router": "18.0.1",
+    "@angular/service-worker": "18.0.1",
     "date-fns": "^3.2.0",
     "lodash-es": "~4.17.21",
     "rxjs": "~7.8.0",
@@ -26,13 +26,13 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.3.8",
-    "@angular/cli": "17.3.8",
-    "@angular/compiler-cli": "17.3.10",
-    "@angular/language-service": "17.3.10",
+    "@angular/build": "^18.0.2",
+    "@angular/cli": "18.0.2",
+    "@angular/compiler-cli": "18.0.1",
+    "@angular/language-service": "18.0.1",
     "@types/lodash-es": "~4.17.5",
     "@types/node": "~20.12.0",
     "ts-node": "~10.9.0",
-    "typescript": "~5.4.0"
+    "typescript": "~5.4.5"
   }
 }
